added admin check for admin routes

+++ src/admin/admin-job/admin-job.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { ReturnModelType } from '@typegoose/typegoose';
 import { InjectModel } from 'nestjs-typegoose';
 import { SerializeService } from '../../../libs/utils/src/serializer/serialize.service';
@@ -7,6 +7,7 @@ import { JobCompletedEntity } from '../../job/entities/job-completed.entity';
 import { JobRequestEntity } from '../../job/entities/job-request.entity';
 import { UserEntity } from '../../user/entities/user.entity';
 import { WalletService } from '../../wallet/wallet.service';
+import { AdminAuthService } from '../admin-auth/auth.service';
 import { CreateAdminJobDto, JobDto } from './dto/admin-job.dto';
 import { JobEntity } from './entities/admin-job.entity';
 
@@ -18,23 +19,30 @@ export class AdminJobService extends SerializeService<JobEntity> {
         @InjectModel(JobCompletedEntity) private readonly completeJobModel: ReturnModelType<typeof JobCompletedEntity>,
         @InjectModel(UserEntity) private readonly userModel: ReturnModelType<typeof UserEntity>,
         private readonly walletService: WalletService,
+        private readonly adminAuthService: AdminAuthService,
     ) {
         super(JobEntity);
     }
 
-    async createJob(body: CreateAdminJobDto) {
+    async createJob(userId: string, body: CreateAdminJobDto) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
+
         const doc = await this.jobModel.create({ ...body, isAvailable: true, isDeleted: false, assignedTo: [] });
 
         return this.toJSON(doc, JobDto);
     }
 
-    async findJobs() {
+    async findJobs(userId: string) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
+
         const docs = await this.jobModel.find({ isAvailable: true, isDeleted: false });
 
         return this.toJSONs(docs, JobDto);
     }
 
-    async approveJobRequest(jobRequestId: string) {
+    async approveJobRequest(userId: string, jobRequestId: string) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
+
         const doc = await this.jobRequestModel.findByIdAndUpdate(jobRequestId, { isAssigned: true }, { new: true });
 
         const job = await this.jobModel.findByIdAndUpdate(
@@ -46,7 +54,9 @@ export class AdminJobService extends SerializeService<JobEntity> {
         return this.toJSON(doc, JobRequestDto);
     }
 
-    async findJobRequests() {
+    async findJobRequests(userId: string) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
+
         const docs = await this.jobRequestModel
             .find({ isAssigned: false, isDeleted: false })
             .populate('job')
@@ -56,21 +66,26 @@ export class AdminJobService extends SerializeService<JobEntity> {
         return this.toJSONs(docs, JobRequestDto);
     }
 
-    async findJobRequestsForCompletion() {
+    async findJobRequestsForCompletion(userId: string) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
+
         const docs = await this.jobRequestModel.find({ isDeleted: false }).populate('job').populate('requester');
         if (!docs) throw new NotFoundException('No job request is found');
 
         return this.toJSONs(docs, JobRequestDto);
     }
 
-    async declineJobRequest(jobRequestId: string) {
+    async declineJobRequest(userId: string, jobRequestId: string) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
+
         const doc = await this.jobRequestModel.findByIdAndDelete(jobRequestId);
 
         return true;
     }
 
-    async deleteOneJob(id: string) {
-        console.log('hit deleteOneJob');
+    async deleteOneJob(userId: string, id: string) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
+
         const doc = await this.jobModel.findOneAndUpdate(
             { _id: id, isAvailable: true, isDeleted: false },
             { isAvailable: false, isDeleted: true },
@@ -80,8 +95,8 @@ export class AdminJobService extends SerializeService<JobEntity> {
         return this.toJSON(doc, JobDto);
     }
 
-    async completeJob(jobRequestId: string) {
-        console.log('hit completeJob()');
+    async completeJob(userId: string, jobRequestId: string) {
+        if (!(await this.adminAuthService.isAdmin(userId))) throw new BadRequestException('This is for admin');
 
         const jobRequest = await this.jobRequestModel.findByIdAndUpdate(
             jobRequestId,
